requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.1
pip==23.3.1
pydantic==2.5.3
aiomysql==0.2.0
astjsonschema==2.17.1
twilio==8.12.0
